# services/mapi-xbt/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bitcoin - Blockchain Indexer API",
    "description": "This API provides core indexer endpoints with support for Bitcoin metaprotocols by delivering real-time, rollback-protected access to Bitcoin's UTXO data, enabling developers to build responsive and reliable blockchain applications without managing complex infrastructure.\n\n#### Key Features:\n- **Real-Time Data with Rollback Protection:** Ensures data accuracy by handling chain reorganizations gracefully, providing live data without sacrificing integrity.\n- **Comprehensive UTXO Indexing:** Specialized pipelines extract, match, and process on-chain information, including handling rollbacks, to provide accurate and up-to-date data.\n\n#### Key Benefits for Developers:\nBy abstracting the complexities of blockchain data retrieval and processing, the Bitcoin Indexer API empowers developers to focus on building innovative applications with confidence in fast and reliable access to historical chain data.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "v0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/assets/inscriptions/{inscription_id}": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Inscription Info",
        "description": "Delivers information about a specific inscription, including its type, current location (UTXO and address), associated collection, and metadata like size and content preview (if text-based). Supports resolution of any inscription for wallet or explorer views.\nA preview of the content body is given only if its type is `\"text/plain\"`. For the whole content, use the complementary endpoint, namely `/assets/inscriptions/{inscription_id}/content_body`.",
        "operationId": "inscription_info",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "description": "Inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "43cb5e2d66f5af8eb04391ef3d4048edc30a1f9edad594f83609fd7861a0f5e1i0"
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampedInscriptionInfo"
                },
                "example": {
                  "data": {
                    "collection_symbol": "kikalepuppete",
                    "content_body_preview": null,
                    "content_length": 934,
                    "content_type": "text/html",
                    "created_at": 867795,
                    "current_location": {
                      "address": "bc1pyae5wn0hxwestrxg4vmwrw58stwzkhp64le0uulahj8xzp5ruz8s3m6xh7",
                      "script_pubkey": "51202773474df733b3058cc8ab36e1ba8782dc2b5c3aaff2fe73fdbc8e610683e08f",
                      "utxo_sat_offset": 0,
                      "utxo_txid": "471852c0dff23a9d345a4550db8a5ff1a9cd0c1ff8e91f344762fba9a1bf000f",
                      "utxo_vout": 1
                    },
                    "inscription_id": "43cb5e2d66f5af8eb04391ef3d4048edc30a1f9edad594f83609fd7861a0f5e1i0",
                    "inscription_number": 76800550
                  },
                  "last_updated": {
                    "block_hash": "000000000000000000018ebea33e7361bc4059eb6ff35d0a80836717accfc4a9",
                    "block_height": 888637
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/assets/inscriptions/{inscription_id}/activity": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Activity by Inscription",
        "description": "Lists all transactions that have involved the given inscription, starting from its origin (reveal transaction) and including all transfers.",
        "operationId": "activity_by_inscription",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "description": "Inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "6fb976ab49dcec017f1e201e84395983204ae1a7c2abf7ced0a85d692e442799i0"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of transactions per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "The order in which the results are sorted (by block height and tx index in the block)",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "string",
                  "default": "asc",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTxByInscription"
                },
                "example": {
                  "data": [
                    {
                      "from": {
                        "address": "bc1pjzf5qmmzt57mtxgrgh42aazhnzwk7ge59e89dl6rde5zwg8he02q9u4xrr",
                        "input_index": 0,
                        "sat_offset": 0,
                        "script_pubkey": "51209093406f625d3db5990345eaaef457989d6f23342e4e56ff436e682720f7cbd4"
                      },
                      "height": 839418,
                      "to": {
                        "address": "bc1p5u4y8vdhn46adxhfv5scfv4c8myykw6r5uyzlavm42k4wgjewktq7xqcyr",
                        "output_vout": 0,
                        "sat_offset": 0,
                        "script_pubkey": "5120a72a43b1b79d75d69ae9652184b2b83ec84b3b43a7082ff59baaad5722597596"
                      },
                      "tx_hash": "b1ef66c2d1a047cbaa6260b74daac43813924378fe08ef8545da4cb79e8fcf00",
                      "tx_index": 1257,
                      "type": "transfer"
                    },
                    {
                      "from": {
                        "address": "bc1p5u4y8vdhn46adxhfv5scfv4c8myykw6r5uyzlavm42k4wgjewktq7xqcyr",
                        "input_index": 0,
                        "sat_offset": 0,
                        "script_pubkey": "5120a72a43b1b79d75d69ae9652184b2b83ec84b3b43a7082ff59baaad5722597596"
                      },
                      "height": 839876,
                      "to": {
                        "address": "bc1ppth27qnr74qhusy9pmcyeaelgvsfky6qzquv9nf56gqmte59vfhqwkqguh",
                        "output_vout": 0,
                        "sat_offset": 0,
                        "script_pubkey": "51200aeeaf0263f5417e40850ef04cf73f43209b13401038c2cd34d201b5e685626e"
                      },
                      "tx_hash": "47c7260764af2ee17aa584d9c035f2e5429aefd96b8016cfe0e3f0bcf04869a3",
                      "tx_index": 887,
                      "type": "transfer"
                    }
                  ],
                  "last_updated": {
                    "block_hash": "000000000000000000001d3fb5743dcb30e12af7b0cd9c8d95c8b1a1fdd5c8d8",
                    "block_height": 839908
                  },
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/assets/inscriptions/{inscription_id}/content_body": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Content by Inscription ID",
        "description": "Retrieves the content body byte array of an inscription. This endpoint is complementary to the `/assets/inscriptions/{inscription_id}` (Inscription Information) endpoint.",
        "operationId": "content_by_inscription_id",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "description": "Inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7d0a2dd897222913d58fc957b0429526117a0a61c964642fe93b077f328ccec1i0"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of bytes per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string: the offset in the content body, in the form of an integer. Use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedContentBody"
                },
                "example": {
                  "data": {
                    "content_body_page": "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiYw==",
                    "remaining_bytes": 2935,
                    "total_length": 3035
                  },
                  "last_updated": {
                    "block_hash": "00000000000000000000ec10254178fe52253f40c1fad252e892d9aa22ee8fa7",
                    "block_height": 866710
                  },
                  "next_cursor": "AAAAAAAAAGQ"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainTip": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string",
            "description": "The hash of the block",
            "example": "0000000000000000000a7f3b7b6b6e1d9a18db65a3b4a3f4f3bcb2e1f1b2d3e7"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "description": "The height of the block in the blockchain",
            "example": 707000,
            "minimum": 0
          }
        }
      },
...
      "ContentBody": {
        "type": "object",
        "required": [
          "content_body_page",
          "total_length",
          "remaining_bytes"
        ],
        "properties": {
          "content_body_page": {
            "type": "string",
            "description": "Base64-encoded representation of a slice of the inscription content body. All types supported."
          },
          "remaining_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Number of bytes remaining in the inscription content body.",
            "minimum": 0
          },
          "total_length": {
            "type": "integer",
            "format": "int64",
            "description": "Number of bytes in entire inscription content body.",
            "minimum": 0
          }
        }
      },
...
      "FromInscriptionLocation": {
        "type": "object",
        "required": [
          "script_pubkey",
          "input_index",
          "sat_offset"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Address-encoding of the script pubkey at which the input containing the inscription resides.",
            "nullable": true
          },
          "input_index": {
            "type": "integer",
            "format": "int32",
            "description": "Index of the input containing the inscription.",
            "minimum": 0
          },
          "sat_offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset of the inscribed satoshi within the input.",
            "minimum": 0
          },
          "script_pubkey": {
            "type": "string",
            "description": "Script pubkey at which the input containing the inscription resides."
          }
        }
      },
...
      "InscriptionInfo": {
        "type": "object",
        "required": [
          "inscription_id",
          "created_at",
          "current_location",
          "content_length"
        ],
        "properties": {
          "collection_symbol": {
            "type": "string",
            "description": "Symbol of collection that the inscription belongs to.",
            "nullable": true
          },
          "content_body_preview": {
            "type": "string",
            "description": "Preview of inscription content body raw data. Max: 100 bytes.\nSupported types: \"text/plain\", \"text/plain;charset=utf-8\", \"application/json\".",
            "nullable": true
          },
          "content_length": {
            "type": "integer",
            "format": "int64",
            "description": "Length of entire inscription content body bytes array.",
            "minimum": 0
          },
          "content_type": {
            "type": "string",
            "description": "Type of the content body.",
            "nullable": true
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Block height of the reveal transaction.",
            "minimum": 0
          },
          "current_location": {
            "$ref": "#/components/schemas/InscriptionLocation"
          },
          "inscription_id": {
            "type": "string",
            "description": "String representation of the inscription ID, whose first coordinate is the reveal\ntransaction hash, and the second coordinate is the index of inscription in the reveal\ntransaction."
          },
          "inscription_number": {
            "type": "integer",
            "format": "int64",
            "description": "Global inscription number.",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "InscriptionLocation": {
        "type": "object",
        "required": [
          "script_pubkey",
          "utxo_sat_offset",
          "utxo_txid",
          "utxo_vout"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Address-encoding of the script pubkey currenty controlling the inscription.",
            "nullable": true
          },
          "script_pubkey": {
            "type": "string",
            "description": "Script pubkey currently controlling the inscription."
          },
          "utxo_sat_offset": {
            "type": "integer",
            "format": "int64",
            "description": "Inscribed sat offset in the UTxO containing the inscription.",
            "minimum": 0
          },
          "utxo_txid": {
            "type": "string",
            "description": "Transaction ID of the UTxO containing the inscription."
          },
          "utxo_vout": {
            "type": "integer",
            "format": "int32",
            "description": "Output index of the UTxO containing the inscription.",
            "minimum": 0
          }
        }
      },
      "InscriptionTxKind": {
        "type": "string",
        "enum": [
          "inscribe",
          "transfer",
          "inscribe_and_spent_as_fee",
          "spent_as_fee"
        ]
      },
...
      "PaginatedContentBody": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ContentBody"
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PaginatedTxByInscription": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TxByInscription"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TimestampedInscriptionInfo": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InscriptionInfo"
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          }
        }
      },
...
      "TxByInscription": {
        "type": "object",
        "required": [
          "height",
          "tx_index",
          "tx_hash",
          "type",
          "to"
        ],
        "properties": {
          "from": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FromInscriptionLocation"
              }
            ],
            "nullable": true
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "description": "Height of block containing transaction that the inscription took part in.",
            "minimum": 0
          },
          "to": {
            "$ref": "#/components/schemas/ToInscriptionLocation"
          },
          "tx_hash": {
            "type": "string",
            "description": "Transaction hash."
          },
          "tx_index": {
            "type": "integer",
            "format": "int32",
            "description": "Index of transaction (in block) that the inscription took part in.",
            "minimum": 0
          },
          "type": {
            "$ref": "#/components/schemas/InscriptionTxKind"
          }
        }
      },
...
    }
  }
}